changed: replaced some floats by double

+++ include/BALL/NMR/johnsonBovey.h
@@ -1,4 +1,4 @@
-// $Id: johnsonBovey.h,v 1.2 1999/12/28 21:29:22 oliver Exp $
+// $Id: johnsonBovey.h,v 1.3 2000/03/26 12:37:53 oliver Exp $
 
 #ifndef BALL_COMMON_H
 #	include <BALL/common.h>
@@ -78,8 +78,8 @@ namespace BALL
 
 			// private Variablen :
 
-			float			n_;
-			float			radius_;
+			double			n_;
+			double			radius_;
 			Vector3		mittelpunkt_;
 			Vector3		normal_;
 
@@ -88,15 +88,15 @@ namespace BALL
 			String**				asrings_;
 			Residue* 				residue_;
 			PDBAtom* 				patom_;
-			float 					shift_;
+			double 					shift_;
 
 
 			// private Funktionen :
 
-			float rf(float,float,float);	// Carlson elliptical Integral of 1st kind
-			float rd(float,float,float);	// Carlson elliptical Integral of 2nd kind
-			float ellf(float,float);	// Legendre elliptical Integral of 1st kind
-			float elle(float,float);	// Legendre elliptical Integral of 2nd kind
+			double rf(double,double,double);	// Carlson elliptical Integral of 1st kind
+			double rd(double,double,double);	// Carlson elliptical Integral of 2nd kind
+			double ellf(double,double);	// Legendre elliptical Integral of 1st kind
+			double elle(double,double);	// Legendre elliptical Integral of 2nd kind
 	};
 
 	//@}
